refactor: mapper 위치 이동
-고수준이 저수준을 가지는 형태라 각 저수준으로 이동
